+++ app/frontend/src/components/Keyboard/layouts/halo96.ts
@@ -1,12 +1,33 @@
 import type { Keyboard } from '../types'
 
-export const Halo65: Keyboard = {
+export const Halo96: Keyboard = {
   halolight: true,
   backlight: true,
   sidelight: true,
   layout: [
     [
       { code: 'esc', color: 'mint' },
+      { code: 'f1' },
+      { code: 'f2' },
+      { code: 'f3' },
+      { code: 'f4' },
+      { code: 'f5', color: 'dark' },
+      { code: 'f6', color: 'dark' },
+      { code: 'f7', color: 'dark' },
+      { code: 'f8', color: 'dark' },
+      { code: 'f9' },
+      { code: 'f10' },
+      { code: 'f11' },
+      { code: 'f12' },
+      { code: 'prntscr', color: 'dark' },
+      { code: 'del', color: 'dark' },
+      { code: 'home', color: 'dark' },
+      { code: 'end', color: 'dark' },
+      { code: 'pgup', color: 'dark' },
+      { code: 'pgdn', color: 'dark' }
+    ],
+    [
+      { code: '~' },
       { code: '1' },
       { code: '2' },
       { code: '3' },
@@ -20,7 +41,10 @@ export const Halo65: Keyboard = {
       { code: '-' },
       { code: '+' },
       { code: 'backspace', color: 'dark', width: 2 },
-      { code: 'del', color: 'dark' }
+      { code: 'numlk', color: 'dark' },
+      { code: '/', color: 'dark' },
+      { code: '×', color: 'dark' },
+      { code: '-', color: 'dark' }
     ],
     [
       { code: 'tab', color: 'dark', width: 1.5 },
@@ -37,7 +61,10 @@ export const Halo65: Keyboard = {
       { code: '[' },
       { code: ']' },
       { code: '\\', width: 1.5 },
-      { code: 'pgup', color: 'dark' }
+      { code: 'num7' },
+      { code: 'num8' },
+      { code: 'num9' },
+      { code: 'numplus', height: 2, color: 'dark' }
     ],
     [
       { code: 'caps', color: 'dark', width: 1.75 },
@@ -53,7 +80,9 @@ export const Halo65: Keyboard = {
       { code: ';' },
       { code: '\'' },
       { code: 'enter', width: 2.25, color: 'orange' },
-      { code: 'pgdn', color: 'dark' }
+      { code: 'num4' },
+      { code: 'num5' },
+      { code: 'num6' }
     ],
     [
       { code: 'lshift', color: 'dark', width: 2.25 },
@@ -69,7 +98,10 @@ export const Halo65: Keyboard = {
       { code: '/' },
       { code: 'rshift', color: 'dark', width: 1.75 },
       { code: 'up', color: 'dark' },
-      { code: 'end', color: 'dark' }
+      { code: 'num1' },
+      { code: 'num2' },
+      { code: 'num3' },
+      { code: 'numenter', height: 2, color: 'mint' }
     ],
     [
       { code: 'lctrl', color: 'dark', width: 1.25 },
@@ -81,7 +113,9 @@ export const Halo65: Keyboard = {
       { width: 0.5 },
       { code: 'left', color: 'dark' },
       { code: 'down', color: 'dark' },
-      { code: 'right', color: 'dark' }
+      { code: 'right', color: 'dark' },
+      { code: 'num0' },
+      { code: 'numdot' }
     ]
   ]
 }